refactor Vocabulary feature

+++ app/src/main/java/ua/com/andromeda/wordgalaxy/data/dao/CategoryDao.kt
@@ -26,7 +26,7 @@ interface CategoryDao {
         GROUP BY categories.id
         """
     )
-    fun findCategoriesWithWordCountAndCompletedWordsCount(parentCategoryId: Int?): Flow<List<VocabularyCategory>>
+    fun findCategoriesWithWordCountAndCompletedWordsCount(parentCategoryId: Long?): Flow<List<VocabularyCategory>>
 
     @Query(
         """
